+++ tests/doc/trajectory/path.cpp
@@ -1,0 +1,18 @@
+// Chemfiles, a modern library for chemistry file reading and writing
+// Copyright (C) Guillaume Fraux and contributors -- BSD license
+
+#include <catch.hpp>
+#include <chemfiles.hpp>
+using namespace chemfiles;
+
+#undef assert
+#define assert CHECK
+
+TEST_CASE() {
+    // [no-run]
+    // [example]
+    auto trajectory = Trajectory("water.nc");
+
+    assert(trajectory.path() == "water.nc");
+    // [example]
+}